Little refactoring to reduce logic on index.php. Moved RssExpander as a core logic system to lib/Bridge.php
